requirements.txt
google-generativeai==0.4.0
PyPDF2==3.0.1
gdown==5.1.0
